Before Lunch

+++ src/main/java/Models/User.java
@@ -1,14 +1,14 @@
 package Models;
 
 public class User {
-    private long id;
+    private Integer id;
     private String username;
     private String email;
     private String password;
 
-    public User() {}
+    public User(){}
 
-    public User(long id, String username, String email, String password) {
+    public User(Integer id, String username, String email, String password) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -25,7 +25,7 @@ public class User {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
